moved all tests into the test folder
