+++ __test__/user-endpoint.spec.js
@@ -1,7 +1,7 @@
-const makeFakeUser = require('../../__test__/fixtures/user')
-const makeUserDb = require('../../__test__/userdb-stub')
-const makeUserActions = require('../services/user-actions')
-const makeUsersEndpointHandler = require('./user-endpoint')
+const makeFakeUser = require('./fixtures/user')
+const makeUserDb = require('./userdb-stub')
+const makeUserActions = require('../src/services/user-actions')
+const makeUsersEndpointHandler = require('../src/controllers/user-endpoint')
 
 
 describe('users endpoint handler', () => {
